requirements.txt
certifi==2021.5.30
pandas==1.3.3
plotly==5.3.1
pymongo==3.12.0
python-dotenv==0.19.2
streamlit==1.3.0